Count the population of a facility isochrones from cpp
Instead of generating a virtual raster from the cutoff of a region
population raster using the isochrone's geojson, and then running
a naive aggregate-population on it, use a custom cpp.

Cpp aggregate-isochrones-population processes all isochrones for
a single region in one run, thus reading the region raster into
memory just once. It also uses block offsets to read only the
parts of the raster actually needed for the computation.

This requires processing of facility isochrones to be batched
by region.

However, there is a small difference in the result. Since the
performance gain ends up being negligible (due to commit
773a4f6a963ff05279), this change is not to be integrated into
master as-is before checking the cause of the difference.

When 773a4f6a963ff05279 is reverted to consider all facilities
whose isochrone affects a given region (and not just the facilities
inside the region), then this changeset will probably yield an
interesting optimisation.